# docs/swagger.yaml
definitions:
  todolist_internal_domain_todo_valueobject.TodoStatistics:
    properties:
      byPriority:
        additionalProperties:
          type: integer
        type: object
      byStatus:
        additionalProperties:
          type: integer
        type: object
      completedToday:
        type: integer
      completionRate:
        type: number
      dueThisWeek:
        type: integer
      dueToday:
        type: integer
      overdue:
        type: integer
      total:
        type: integer
    type: object
...
  todolist_internal_dto.CreateTodoRequest:
    properties:
      description:
        maxLength: 1000
        type: string
      due_date:
        type: string
      priority:
        enum:
        - low
        - medium
        - high
        - critical
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        maxLength: 200
        minLength: 3
        type: string
    required:
    - priority
    - title
    type: object
...
  todolist_internal_dto.ErrorInfo:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      details:
        additionalProperties: {}
        type: object
      message:
        example: Invalid input data
        type: string
    type: object
  todolist_internal_dto.PaginatedResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/todolist_internal_dto.Pagination'
      success:
        example: true
        type: boolean
    type: object
  todolist_internal_dto.Pagination:
    properties:
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      total:
        example: 100
        type: integer
      total_pages:
        example: 5
        type: integer
    type: object
...
  todolist_internal_dto.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/todolist_internal_dto.ErrorInfo'
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
  todolist_internal_dto.TodoResponse:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: integer
      is_overdue:
        type: boolean
      priority:
        type: string
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  todolist_internal_dto.UpdateTodoRequest:
    properties:
      description:
        maxLength: 1000
        type: string
      due_date:
        type: string
      priority:
        enum:
        - low
        - medium
        - high
        - critical
        type: string
      status:
        enum:
        - pending
        - in_progress
        - completed
        - cancelled
        type: string
      title:
        maxLength: 200
        minLength: 3
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/todos:
    get:
      consumes:
      - application/json
      description: List todos with filters and pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - collectionFormat: csv
        description: Filter by status
        in: query
        items:
          enum:
          - pending
          - in_progress
          - completed
          - cancelled
          type: string
        name: status
        type: array
      - collectionFormat: csv
        description: Filter by priority
        in: query
        items:
          enum:
          - low
          - medium
          - high
          - critical
          type: string
        name: priority
        type: array
      - collectionFormat: csv
        description: Filter by tags
        in: query
        items:
          type: string
        name: tags
        type: array
      - description: Search in title and description
        in: query
        name: search
        type: string
      - description: Filter overdue todos
        in: query
        name: is_overdue
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/todolist_internal_dto.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/todolist_internal_dto.TodoResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/todolist_internal_dto.Response'
      security:
      - BearerAuth: []
      summary: List todos
      tags:
      - todos
    post:
      consumes:
      - application/json
      description: Create a new todo item
      parameters:
      - description: Todo data
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/todolist_internal_dto.CreateTodoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/todolist_internal_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/todolist_internal_dto.TodoResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/todolist_internal_dto.Response'
      security:
      - BearerAuth: []
      summary: Create a new todo
      tags:
      - todos
  /api/v1/todos/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a todo item
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/todolist_internal_dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/todolist_internal_dto.Response'
      security:
      - BearerAuth: []
      summary: Delete todo
      tags:
      - todos
    get:
      consumes:
      - application/json
      description: Get todo details by ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/todolist_internal_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/todolist_internal_dto.TodoResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/todolist_internal_dto.Response'
      security:
      - BearerAuth: []
      summary: Get todo by ID
      tags:
      - todos
    put:
      consumes:
      - application/json
      description: Update todo details
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      - description: Todo data
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/todolist_internal_dto.UpdateTodoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/todolist_internal_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/todolist_internal_dto.TodoResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/todolist_internal_dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/todolist_internal_dto.Response'
      security:
      - BearerAuth: []
      summary: Update todo
      tags:
      - todos
  /api/v1/todos/{id}/complete:
    post:
      consumes:
      - application/json
      description: Mark todo as completed
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/todolist_internal_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/todolist_internal_dto.TodoResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/todolist_internal_dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/todolist_internal_dto.Response'
      security:
      - BearerAuth: []
      summary: Complete todo
      tags:
      - todos
  /api/v1/todos/statistics:
    get:
      consumes:
      - application/json
      description: Get todo statistics for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/todolist_internal_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/todolist_internal_domain_todo_valueobject.TodoStatistics'
              type: object
      security:
      - BearerAuth: []
      summary: Get todo statistics
      tags:
      - todos
swagger: "2.0"